+++ Pintervention/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 1.1.0
+
+  * Changed `LoadPlayerNames()` to run as `ZNetScene.Awake()` postfix.
+  * Added feature to save filtered pin owners between sessions.
+
 ### 1.0.0
 
   * Initial release.